Major refactoring
Refactored the SpecBuilder module to better decouple the model from the view.
Refactored the Algorithm modules to decouple them from eachother.
Removed all Algorithms except two of them.
Removed the parameters for elastic scattering and angular spread.
Re-named the inelastic decay parameter to normalization parameter.
The format of the JSON files for scatteres is now different.