Added Spectrum Lineshape Buffer
In order to speed up the refresh rate for when someone is editing a synthetic lineshape, where there are lots of components, I made a spectrum lineshape bufffer, which holds all the components y-values in a numpy array.
The interface with the two kinds of syntheitc spectra (i.e. LossFunction and Synthetic input spectrum) needed to be modified a bit.